continue to clean
