+++ CFQ/Data/LargeButtonData.swift
@@ -1,0 +1,8 @@
+
+import SwiftUI
+
+struct LargeButtonData {
+    let background: Color
+    let foregroundColor: Color
+    let hasStoke: Bool
+}
