+++ CFQ/Data/MoodData.swift
@@ -1,0 +1,12 @@
+
+import SwiftUI
+
+class MoodData {
+    let icon: UIImage
+    let title: String
+    
+    init(icon: UIImage, title: String) {
+        self.icon = icon
+        self.title = title
+    }
+}
