+++ CFQ/Utils/Assets.xcassets/Colors/Black/blackMain.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "19",
-          "green" : "17",
-          "red" : "17"
+          "blue" : "0x13",
+          "green" : "0x11",
+          "red" : "0x11"
         }
       },
       "idiom" : "universal"
