+++ CFQ/Utils/Assets.xcassets/Colors/Black/blackSurface.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "30",
-          "green" : "30",
-          "red" : "30"
+          "blue" : "0x1E",
+          "green" : "0x1E",
+          "red" : "0x1E"
         }
       },
       "idiom" : "universal"
