+++ CFQ/Utils/Assets.xcassets/Colors/Other/Inactive.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "255",
-          "green" : "71",
-          "red" : "151"
+          "blue" : "0xFF",
+          "green" : "0x47",
+          "red" : "0x97"
         }
       },
       "idiom" : "universal"
