+++ CFQ/Utils/Assets.xcassets/Colors/Other/grayCard.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0xE6",
-          "green" : "0xFF",
-          "red" : "0x47"
+          "blue" : "0x69",
+          "green" : "0x65",
+          "red" : "0x4B"
         }
       },
       "idiom" : "universal"
@@ -23,9 +23,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "230",
-          "green" : "255",
-          "red" : "71"
+          "blue" : "0x69",
+          "green" : "0x65",
+          "red" : "0x4B"
         }
       },
       "idiom" : "universal"
