+++ CFQ/Utils/Assets.xcassets/Colors/White/whitePrimary.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "250",
-          "green" : "249",
-          "red" : "248"
+          "blue" : "0xFA",
+          "green" : "0xF9",
+          "red" : "0xF8"
         }
       },
       "idiom" : "universal"
