+++ CFQ/Utils/Assets.xcassets/Colors/purple/purpleDark.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "211",
-          "green" : "63",
-          "red" : "93"
+          "blue" : "0xD3",
+          "green" : "0x3F",
+          "red" : "0x5D"
         }
       },
       "idiom" : "universal"
