+++ CFQ/Utils/Assets.xcassets/Colors/purple/purpleLight.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "227",
-          "green" : "95",
-          "red" : "123"
+          "blue" : "0xE3",
+          "green" : "0x5F",
+          "red" : "0x7B"
         }
       },
       "idiom" : "universal"
