+++ CFQ/Utils/Enum/Moods.swift
@@ -38,21 +38,3 @@ struct Mood {
         }
     }
 }
-
-class MoodData {
-    let icon: UIImage
-    let title: String
-    
-    init(icon: UIImage, title: String) {
-        self.icon = icon
-        self.title = title
-    }
-}
-
-
-#Preview {
-    ZStack {
-        Color.blue.ignoresSafeArea()
-        MoodTemplateView()
-    }
-}
