+++ CFQ/Widget/Atome/Button/LargeButtonView.swift
@@ -1,43 +1,6 @@
 
 import SwiftUI
 
-
-enum LargeButtonType {
-    case signNext
-    case signBack
-    case teamCreate
-    case addParticipant
-    
-    var data: LargeButtonData {
-        switch self {
-        case .signNext, .addParticipant:
-            return LargeButtonData(
-                background: .black,
-                foregroundColor: .white,
-                hasStoke: true
-            )
-        case .signBack:
-            return LargeButtonData(
-                background: .clear,
-                foregroundColor: .purple,
-                hasStoke: false
-            )
-        case .teamCreate:
-            return LargeButtonData(
-                background: .white,
-                foregroundColor: .black,
-                hasStoke: false
-            )
-        }
-    }
-}
-
-struct LargeButtonData {
-    let background: Color
-    let foregroundColor: Color
-    let hasStoke: Bool
-}
-
 struct LargeButtonView: View{
     var action: () -> Void
     var title: String
