+++ CFQ/Widget/Atome/GradientCard.swift
@@ -6,9 +6,9 @@ struct GradientCardView: View {
         VStack {
             LinearGradient(
                 gradient: Gradient(stops: [
-                    Gradient.Stop(color: Color(hex: "111113"), location: 0.0),
-                    Gradient.Stop(color: Color(hex: "111113"), location: 0.4),
-                    Gradient.Stop(color: Color(hex: "4B6569"), location: 1.0)
+                    Gradient.Stop(color: Color(.blackMain), location: 0.0),
+                    Gradient.Stop(color: Color(.blackMain), location: 0.4),
+                    Gradient.Stop(color: Color(.grayCard), location: 1.0)
                 ]),
                 startPoint: .bottom,
                 endPoint: .topTrailing
